feat: 초기 보스 리젠, 리스폰 관련 기능 & UI 제작
1. 보스 리젠, 리스폰 관련 기능 및 UI 구현 완료

+++ Cat_Merge/Assets/1.Scripts/Contents/Mouse.cs
@@ -13,11 +13,11 @@ public class Mouse : MonoBehaviour
     private int mouseGrade;             // Áã µî±Þ (½ºÅ×ÀÌÁö)
     public int MouseGrade { get => mouseGrade; set => mouseGrade = value; }
 
-    private int mouseDamage;            // Áã ÀüÅõ·Â
-    public int MouseDamage { get => mouseDamage; set => mouseDamage = value; }
+    private float mouseDamage;            // Áã ÀüÅõ·Â
+    public float MouseDamage { get => mouseDamage; set => mouseDamage = value; }
 
-    private int mouseHp;                // Áã Ã¼·Â
-    public int MouseHp { get => mouseHp; set => mouseHp = value; }
+    private float mouseHp;                // Áã Ã¼·Â
+    public float MouseHp { get => mouseHp; set => mouseHp = value; }
 
     private Sprite mouseImage;          // Áã ÀÌ¹ÌÁö
     public Sprite MouseImage { get => mouseImage; set => mouseImage = value; }
@@ -25,7 +25,7 @@ public class Mouse : MonoBehaviour
     private int numOfAttack;            // °ø°Ý Å¸°Ù ¼ö
     public int NumOfAttack { get => numOfAttack; set => numOfAttack = value; }
 
-    public Mouse(int mouseId, string mouseName, int mouseGrade, int mouseDamage, int mouseHp, Sprite mouseImage, int numOfAttack)
+    public Mouse(int mouseId, string mouseName, int mouseGrade, float mouseDamage, float mouseHp, Sprite mouseImage, int numOfAttack)
     {
         MouseId = mouseId;
         MouseName = mouseName;
